requirements.txt
# ui framework for web interface
streamlit==1.29.0
# llm api client for openai models
openai>=1.14.0
# llm api client for anthropic models
anthropic>=0.34.0
# baseline translation library
deep-translator==1.11.4
# data manipulation and csv handling
pandas==2.2.2
# web scraping with browser automation
selenium==4.23.1
# http requests for scraping
requests==2.32.3
# html parsing for scraping
beautifulsoup4==4.12.3
# environment variable management
python-dotenv==1.0.1
# chrome driver management
webdriver-manager==4.0.2
# nlp tools for evaluation metrics
nltk==3.9.1
# ssl certificate handling on windows
python-certifi-win32==1.6.1
# stealth browser automation
undetected-chromedriver==3.5.5
# bypass cloudflare protection
cloudscraper==1.2.71